requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

python_dateutil == 2.8.2
python_dotenv == 0.20.0
requests == 2.27.1
rich == 12.5.1
